Refactor SwapHistory filters and improve filtering logic
- Updated Filters component to separate NoMatches and SearchResult into individual files for better organization.
- Enhanced filtering logic in filterSwaps to utilize Address equality checks for wallet addresses.
- Modified useHistoryFilters to remove unnecessary networks parameter and streamline wallet address handling.
- Changed role attribute in CheckboxRow from 'button' to 'checkbox' for improved accessibility.
