+++ components/SwapHistory/Filters/types.ts
@@ -1,3 +1,5 @@
+import { Wallet } from '@/Models/WalletProvider'
+
 export type FilterNetworkOption = {
     name: string
     display_name: string
@@ -8,3 +10,5 @@ export type FilterOpts = {
     walletAddrs: string[] | null
     networks: string[] | null
 }
+
+export const walletIdOf = (w: Wallet): string => w.internalId ?? w.address
